Consolidate all files SWR hooks in one file
