+++ front/components/data_source/MultipleDocumentsUpload.tsx
@@ -17,7 +17,7 @@ import type {
   FileBlobWithFileId,
 } from "@app/hooks/useFileUploaderService";
 import { useFileUploaderService } from "@app/hooks/useFileUploaderService";
-import { useUpsertFileAsDatasourceEntry } from "@app/lib/swr/file";
+import { useUpsertFileAsDatasourceEntry } from "@app/lib/swr/files";
 import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import type {
   DataSourceViewType,
